pyproject.toml
[project]
name = "healthbench"
version = "0.1.0"
description = "Implementation of OpenAI's HealthBench evaluation framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=2.0.0",
    "tqdm>=4.0.0",
    "tabulate>=0.9.0",
    "matplotlib>=3.0.0",
    "seaborn>=0.13.0",
    "openai>=2.0.0",
    "anthropic>=0.75.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "blobfile>=3.0.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
]